Revert "Fix unified definition navigation and Ctrl+Click local resolution (#32)" (#35)
This reverts commit dd066d51c69cdb092a394285c4e7aabd482b0c7c.
