+++ package.json
@@ -539,6 +539,11 @@
       ],
       "editor/title": [
         {
+          "command": "vscode-objectscript.ccs.goToDefinition",
+          "group": "navigation@0",
+          "when": "editorTextFocus && editorLangId =~ /^objectscript/"
+        },
+        {
           "command": "-editor.action.revealDefinition",
           "group": "navigation@0",
           "when": "editorLangId =~ /^objectscript/"
